requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
